Rename Generator to DataGenerator
