+++ lib/data_generator.rb
